Feat: clean sundries (#18)
* chore: remove useless mumbai network support

* chore: coverage and format script, remove docgen and linter

* chore: renaming

* feat: lint tool

* chore: use pure solhint instead of hardhat-solhint

* chore: use single quotes

* chore: use explicit type in seacows-periphery contracts

* chore: use explicit type in seacows-amm contracts

* chore: solhint fix

* chore: solhint fix

* feat: use custom errors required by solhint

* delete: interface ISeacowsCallback

* feat: slither.config file

* feat: hardhat clean command

+++ packages/seacows-amm/contracts/base/SeacowsRewarder.sol
@@ -1,14 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0
 pragma solidity ^0.8.13;
 
-import {IERC3525Receiver} from '@solvprotocol/erc-3525/IERC3525Receiver.sol';
-import '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import '@openzeppelin/contracts/token/ERC721/IERC721Receiver.sol';
-import '@openzeppelin/contracts/utils/introspection/ERC165.sol';
-
-import {ERC3525} from '@solvprotocol/erc-3525/ERC3525.sol';
-import {IERC3525} from '@solvprotocol/erc-3525/IERC3525.sol';
-import {ISeacowsERC3525} from '../interfaces/ISeacowsERC3525.sol';
+import {IERC20} from '@openzeppelin/contracts/token/ERC20/IERC20.sol';
 import {ISeacowsRewarder} from '../interfaces/ISeacowsRewarder.sol';
 import {SeacowsPairMetadata} from './SeacowsPairMetadata.sol';
 
@@ -28,10 +21,10 @@ contract SeacowsRewarder is ISeacowsRewarder, SeacowsPairMetadata {
         uint256 unclaimedFee;
     }
 
-    function getPendingFee(uint _tokenId) public view virtual returns (uint256) {
+    function getPendingFee(uint256 _tokenId) public view virtual returns (uint256) {
         PositionInfo storage _info = positionInfos[_tokenId];
 
-        uint _accRewardTokenPerShare = accRewardPerShare;
+        uint256 _accRewardTokenPerShare = accRewardPerShare;
         if (feeBalance != lastFeeBalance) {
             _accRewardTokenPerShare += ((feeBalance - lastFeeBalance) * ACC_REWARD_PER_SHARE_PRECISION) / totalSupply();
         }
@@ -54,9 +47,9 @@ contract SeacowsRewarder is ISeacowsRewarder, SeacowsPairMetadata {
         lastFeeBalance = feeBalance;
     }
 
-    function updatePositionFee(uint tokenId) public onlyPositionManager {
+    function updatePositionFee(uint256 tokenId) public onlyPositionManager {
         PositionInfo storage info = positionInfos[tokenId];
-        uint liquidity = balanceOf(tokenId);
+        uint256 liquidity = balanceOf(tokenId);
 
         if (liquidity != 0) {
             uint256 _pending = (liquidity * accRewardPerShare) / ACC_REWARD_PER_SHARE_PRECISION - info.feeDebt;
@@ -66,9 +59,9 @@ contract SeacowsRewarder is ISeacowsRewarder, SeacowsPairMetadata {
         }
     }
 
-    function updatePositionFeeDebt(uint tokenId) public onlyPositionManager {
+    function updatePositionFeeDebt(uint256 tokenId) public onlyPositionManager {
         PositionInfo storage info = positionInfos[tokenId];
-        uint liquidity = balanceOf(tokenId);
+        uint256 liquidity = balanceOf(tokenId);
         info.feeDebt = (liquidity * accRewardPerShare) / ACC_REWARD_PER_SHARE_PRECISION;
     }
 
@@ -76,7 +69,7 @@ contract SeacowsRewarder is ISeacowsRewarder, SeacowsPairMetadata {
     @notice Collect Swap fee for Position NFT
     @param _tokenId The Position NFT ID contract address
    */
-    function collect(uint256 _tokenId) public returns (uint _fee) {
+    function collect(uint256 _tokenId) public returns (uint256 _fee) {
         updateSwapFee();
         uint256 _feeAmount = getPendingFee(_tokenId);
         positionInfos[_tokenId].unclaimedFee = 0;
