+++ packages/seacows-amm/contracts/interfaces/ISeacowsPairMetadata.sol
@@ -6,6 +6,7 @@ interface ISeacowsPairMetadata {
 
     function collection() external view returns (address);
 
+    /* solhint-disable func-name-mixedcase */
     function PERCENTAGE_PRECISION() external view returns (uint64);
 
     function ONE_PERCENT() external view returns (uint64);
@@ -14,9 +15,11 @@ interface ISeacowsPairMetadata {
 
     function MAX_PROTOCOL_FEE_PERCENT() external view returns (uint64);
 
+    /* solhint-enable func-name-mixedcase */
+
     function totalSupply() external view returns (uint256);
 
-    function balanceOf(uint _tokenId) external view returns (uint256);
+    function balanceOf(uint256 _tokenId) external view returns (uint256);
 
-    function ownerOf(uint _tokenId) external view returns (address);
+    function ownerOf(uint256 _tokenId) external view returns (address);
 }
