+++ packages/seacows-periphery/contracts/interfaces/ISeacowsSwapCallback.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.13;
 
 interface ISeacowsSwapCallback {
-    function seacowsSwapCallback(bytes calldata data) external returns (uint tokenAmountIn, uint[] memory idsIn);
+    function seacowsSwapCallback(bytes calldata data) external returns (uint256 tokenAmountIn, uint256[] memory idsIn);
 }
