+++ packages/seacows-periphery/contracts/interfaces/IWETH.sol
@@ -10,7 +10,7 @@ interface IWETH {
 
     function transfer(address to, uint256 value) external returns (bool);
 
-    function transferFrom(address src, address dst, uint wad) external returns (bool);
+    function transferFrom(address src, address dst, uint256 wad) external returns (bool);
 
     function approve(address guy, uint256 wad) external returns (bool);
